Read files from my MSigDB cache

Reads an MSigDB gmt file and returns a list of genes in pathways.

read_msigdb(
  collection = "h.all",
  pathways = NULL,
  version = "7.4",
  identifier = "symbols",
  dbroot = "/camp/stp/babs/working/barrinc/db/msigdb"
)

Arguments

collection

Name of the collection

pathways

Character vector of pathways in collection to keep

version

Release version

identifier

Identifier type: symbols or entrez (entrez may not exist!!)

dbroot

Path to the root of the MSigDB files

Value

A list of genes in pathways, with pathway name as the key and a character vector of gene identifiers (identifer).