Knit an R Markdown File and Export to Google Drive

This function allows you to knit a specified R Markdown file locally and export it to the Google Drive folder for which you provided a link. NOTE that if you have not used googledrive::drive_auth this will prompt you to authorize a Google account in a new browser tab. If you do not check the box in that screen before continuing you will not be able to use this function until you clear your browser cache and re-authenticate. I recommend invoking drive_auth beforehand to reduce the chances of this error

rmd_export(
  rmd = NULL,
  out_path = getwd(),
  out_name = NULL,
  out_type = "html",
  drive_link
)

Arguments

rmd

(character) name and path to R markdown file to knit

out_path

(character) path to the knit file's destination (defaults to path returned by getwd)

out_name

(character) desired name for knit file (with or without file suffix)

out_type

(character) either "html" or "pdf" depending on what YML entry you have in the output: field of your R Markdown file

drive_link

(character) full URL of drive folder to upload the knit document

Examples

# Authorize R to interact with GoogleDrive
# googledrive::drive_auth()
## NOTE: See above warning about possible misstep at this stage

# Use `rmd_export()` to knit and export an .Rmd file
# rmd_export(rmd = "my_markdown.Rmd",
# in_path = file.path("Folder in my WD with the .Rmd named in `rmd`"),
# out_path = file.path("Folder in my WD to save the knit file to"),
# out_name = "desired name for output",
# out_type = "html",
# drive_link = "https://...Google Drive link...")