Read data from a cool (or mcool) file

Reads the cool h5 file to provide the whole interaction matrix or a subset of coordinates.

get_cool_interaction_matrix(cool, mcool, resolution)

get_cool_interaction_submatrix(
  chrom1,
  pos1,
  range1,
  chrom2 = chrom1,
  pos2 = pos1,
  range2 = range1,
  mcool,
  resolution
)

Arguments

cool, mcool

Path to the h5 file

mcool, resolution

Path to h5 file and resolution to extract

resolution

Bin size in the mcool to extract

chrom1, chrom2

Chromosomes for intervals

pos1, pos2

Start coordinate of intervals

range1, range2

Size of intervals (base pairs)

Value

A data.frame of 2D genomic coordinates with their weights and balanced score.

Details

The cool path takes priority and if mcool is provided so too must resolution.

Coordinates are filtered from the whole matrix via get_cool_interaction_matrix().

Functions

get_cool_interaction_matrix(): Get a genome-wide interaction matrix

get_cool_interaction_submatrix(): Get an interaction matrix for a pair of genomic regions